Added UMP SDK integration to banner example (#1083)

+++ samples/admob/banner_example/lib/main.dart
@@ -3,9 +3,10 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
 
+import 'consent_manager.dart';
+
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
-  MobileAds.instance.initialize();
   runApp(const MaterialApp(
     home: BannerExample(),
   ));
@@ -20,6 +21,10 @@ class BannerExample extends StatefulWidget {
 }
 
 class BannerExampleState extends State<BannerExample> {
+  static const privacySettingsText = 'Privacy Settings';
+
+  final _consentManager = ConsentManager();
+  var _isMobileAdsInitializeCalled = false;
   BannerAd? _bannerAd;
   bool _isLoaded = false;
   Orientation? _currentOrientation;
@@ -29,13 +34,34 @@ class BannerExampleState extends State<BannerExample> {
       : 'ca-app-pub-3940256099942544/2435281174';
 
   @override
+  void initState() {
+    super.initState();
+
+    _consentManager.gatherConsent((consentGatheringError) {
+      if (consentGatheringError != null) {
+        // Consent not obtained in current session.
+        debugPrint(
+            "${consentGatheringError.errorCode}: ${consentGatheringError.message}");
+      }
+
+      // Attempt to initialize the Mobile Ads SDK.
+      _initializeMobileAdsSDK();
+    });
+
+    // This sample attempts to load ads using consent obtained in the previous session.
+    _initializeMobileAdsSDK();
+  }
+
+  @override
   Widget build(BuildContext context) {
     return MaterialApp(
         title: 'Banner Example',
         home: Scaffold(
             appBar: AppBar(
-              title: const Text('Banner Example'),
-            ),
+                title: const Text('Banner Example'),
+                actions: _isMobileAdsInitializeCalled
+                    ? _privacySettingsAppBarAction()
+                    : null),
             body: OrientationBuilder(
               builder: (context, orientation) {
                 if (_currentOrientation != orientation) {
@@ -43,7 +69,6 @@ class BannerExampleState extends State<BannerExample> {
                   _loadAd();
                   _currentOrientation = orientation;
                 }
-
                 return Stack(
                   children: [
                     if (_bannerAd != null && _isLoaded)
@@ -63,10 +88,52 @@ class BannerExampleState extends State<BannerExample> {
             )));
   }
 
+  List<Widget> _privacySettingsAppBarAction() {
+    return <Widget>[
+      // Regenerate the options menu to include a privacy setting.
+      FutureBuilder(
+          future: _consentManager.isPrivacyOptionsRequired(),
+          builder: (context, snapshot) {
+            final bool visibility = snapshot.data ?? false;
+            return Visibility(
+                visible: visibility,
+                child: PopupMenuButton<String>(
+                  onSelected: (String result) {
+                    if (result == privacySettingsText) {
+                      _consentManager.showPrivacyOptionsForm((formError) {
+                        if (formError != null) {
+                          debugPrint(
+                              "${formError.errorCode}: ${formError.message}");
+                        }
+                      });
+                    }
+                  },
+                  itemBuilder: (BuildContext context) =>
+                      <PopupMenuEntry<String>>[
+                    const PopupMenuItem<String>(
+                        value: privacySettingsText,
+                        child: Text(privacySettingsText))
+                  ],
+                ));
+          })
+    ];
+  }
+
   /// Loads and shows a banner ad.
   ///
   /// Dimensions of the ad are determined by the width of the screen.
   void _loadAd() async {
+    // Only load an ad if the Mobile Ads SDK has gathered consent aligned with
+    // the app's configured messages.
+    var canRequestAds = await _consentManager.canRequestAds();
+    if (!canRequestAds) {
+      return;
+    }
+
+    if (!mounted) {
+      return;
+    }
+
     // Get an AnchoredAdaptiveBannerAdSize before loading the ad.
     final size = await AdSize.getCurrentOrientationAnchoredAdaptiveBannerAdSize(
         MediaQuery.sizeOf(context).width.truncate());
@@ -100,6 +167,26 @@ class BannerExampleState extends State<BannerExample> {
         onAdImpression: (Ad ad) {},
       ),
     ).load();
+  }
+
+  /// Initialize the Mobile Ads SDK if the SDK has gathered consent aligned with
+  /// the app's configured messages.
+  void _initializeMobileAdsSDK() async {
+    if (_isMobileAdsInitializeCalled) {
+      return;
+    }
+
+    var canRequestAds = await _consentManager.canRequestAds();
+    if (canRequestAds) {
+      setState(() {
+        _isMobileAdsInitializeCalled = true;
+      });
+
+      // Initialize the Mobile Ads SDK.
+      MobileAds.instance.initialize();
+      // Load an ad.
+      _loadAd();
+    }
   }
 
   @override
